Added a check for a weird bug that started happening. I'm not sure if this has to do with my virtual desktop program (which has caused issues with other programs before) or what.
I'll briefly outline the issue:
For whatever reason, if the process needs to restart in non-headless mode (due to a very specific bug) it sometimes can't find aria2c (despite it being in the same folder and the working directory being set correctly).

Shrug? If I get any bug reports about this I might be able to figure it out, until then this will have to do. The non-headless bug is in in of itself a really rare one, so I'm hoping this won't be a huge issue.